+++ src/webviews/components/issue/view-issue-screen/JiraIssuePage.tsx
@@ -479,7 +479,6 @@ export default class JiraIssuePage extends AbstractIssueEditorPage<Emit, Accept,
     };
 
     handleRefresh = () => {
-        this.setState({ isSomethingLoading: true, loadingField: 'refresh' });
         this.postMessage({ action: 'refreshIssue' });
     };