+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bug Fixes
 
 - Removed setting options that had no practical or business logic use.
+- Added loading indicator to Refresh button for Jira issue auto-refreshes
 - Fixed a bug where Jira site icon is broken in create Jira issue page
 - Fixed a bug with getting the error when deleting multiple instances
 - Fixed a bug where we could not assign people to a Jira issue as DC users
@@ -26,7 +27,6 @@
 
 - Fixed a bug where issues would not render due to malformed epic fields (#665)
 - Fixed status button in smaller screen size
--
 
 ## What's new in 3.8.7